Random sampling via Ulrich's model, clearing targets determined endongenously

+++ simulator/src/main/java/ca/ubc/cs/beta/fcc/simulator/station/AStationInfoDecorator.java
@@ -23,7 +23,7 @@ public abstract class AStationInfoDecorator implements IStationInfo {
     }
 
     @Override
-    public Double getVolume() {
+    public Integer getVolume() {
         return decorated.getVolume();
     }
